Use a shared odr system group
To securely share the PAD FIFO between the odr-padenc and odr-audioenc
services a dedicated group named `odr` is used. This allows odr-padenc
to write to the FIFO it owns, and odr-audioenc to read from the FIFO.

Yes, this breaks backwards compatibility, but I couldn't find an other
simple solution. As far as I know, there is no odr-padenc RPM package in
production, so it shouldn't generate too much noise right now. For
systems which do already make use of the odr-padenc systemd service
unit, the following steps are necessary before updating:
1. Stop all odr-padenc services:
   systemctl stop odr-padenc@<INSTANCE>.service
2. Remove the temporary directory
   rm -rf /var/tmp/odr/padenc
3. Delete the user and group
   userdel odr-padenc
   groupdel odr-padenc
4. Update the rpm